Fixing the joining tables issue and doing other stuff that i dont remember

+++ MovieAPI/src/main/java/com/encora/movieapi/repositories/UserRepository.java
@@ -6,7 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.encora.movieapi.Entities.User;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
-
-
+public interface UserRepository extends JpaRepository<User, Long> {
+    
 }
